pyproject.toml
[project]
name = "shopify-square-connect"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
 "shopifyapi>=12.7.0",
 "squareup>=41.0.0.20250319",
]
